Implement Serde for Fp2
The serde implementation sets the `Fp2` as 2 - 48 bytes
`Fp`s and then encodes them and does the decoding and decompression
on the deserialization.

The deserialization will fail if any of the `Fp`s was encoded
incorrectly or the bytes are less than 48 for each of them.

Added tests for  roundtrip de/serialization.

Task (2/3) to close #13