Implement Serde for Fp
The serde implementation sets the `Fp` as 48 bytes,
then encodes it and does the decoding and decompression
on the deserialization.

The deserialization will fail if the `Fp` was encoded
incorrectly or the bytes are less than 48.

Added tests for G1Affine roundtrip de/serialization.

We implement this serde support for `Fp` since it will
be necessary so later is easier for us to de/serialize
`G2Prepared` which does not carry any `to_bytes` impl.