Implement Serde for Fp2
The serde implementation sets the `Fp2` as 2 - 48 bytes
`Fp`s and then encodes them and does the decoding and decompression
on the deserialization.

The deserialization will fail if any of the `Fp`s was encoded
incorrectly or the bytes are less than 48 for each of them.

Added tests for roundtrip de/serialization.

We implement this serde support for `Fp2` since it will
be necessary so later is easier for us to de/serialize
`G2Prepared` which does not carry any `to_bytes` impl.